Migrate from getopt to click
This commit will migrate the script from getopt to the click module. As well as reducing script size from 122 lines -> 84, Click provides a cleaner and more extensible way to interact via the CLI.

The commit also adds the ability to pass in the path to a config file that sets the hash_fields, which feels a lot cleaner than passing in via CLI and provides a mean to peer review columns via source control. The config file also enables the consumer to specify and out_dir, in the case that the consumer needs to customise the output directory.

Lastly, this commit builds in functionality to get the filename from the source path defined by the consumer.