Rename hash-csv to hash_csv and add CI GitHub Action Workflow
Renamed hash-csv to hash_csv to lay the groundwork for importing the module in other scripts (i.e. for testing).

Added some basic pytest tests - these need to be refined so that they're useful.

Added a GitHub Actions Workflow to attempt to perform CI of the script on commit to a PR.